feat: 게시글 수정/삭제 시 작성자 본인 확인 로직 구현

+++ src/main/java/com/newspeed/sixteenflow/domain/post/service/PostServiceImpl.java
@@ -11,6 +11,7 @@ import com.newspeed.sixteenflow.domain.post.dto.update.UpdatePostResponseDto;
 import com.newspeed.sixteenflow.domain.post.entity.Post;
 import com.newspeed.sixteenflow.domain.post.repository.PostRepository;
 import com.newspeed.sixteenflow.global.exception.post.PostNotFoundException;
+import com.newspeed.sixteenflow.global.exception.post.PostUnauthorizedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -29,10 +30,7 @@ public class PostServiceImpl implements PostService {
     @Transactional
     @Override
     public CreatePostResponseDto createPost(Long memberId, CreatePostRequestDto requestDto) {
-        // FIXME: 현재는 임시로 ResponseStatusException 사용 중.
-        // 추후 member 도메인에서 커스텀 예외(MemberNotFoundException 등) 정의되면 교체할 것.
-        Member findMember = memberRepository.findById(memberId)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Member findMember = findMemberByIdOrElseThrow(memberId);
 
         Post post = new Post(
                 requestDto.getContent(),
@@ -55,8 +53,7 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public PostResponseDto findPostById(Long postId) {
-        Post findPost = postRepository.findById(postId)
-                .orElseThrow(() -> new PostNotFoundException());
+        Post findPost = findPostByIdOrElseThrow(postId);
 
         // TODO: 현재는 좋아요 수(postLike)와 댓글 수(postComment)를 null로 설정하고 있음.
         // 추후 PostLike, Comment 데이터와 연동하여 실제 수치를 계산해 넣을 예정.
@@ -65,11 +62,11 @@ public class PostServiceImpl implements PostService {
 
     @Transactional
     @Override
-    public UpdatePostResponseDto updatePost(Long postId, UpdatePostRequestDto requestDto) {
-        Post findPost = postRepository.findById(postId)
-                .orElseThrow(() -> new PostNotFoundException());
+    public UpdatePostResponseDto updatePost(Long memberId, Long postId, UpdatePostRequestDto requestDto) {
+        Post findPost = findPostByIdOrElseThrow(postId);
 
         // TODO: 추후에 작성자를 확인하는 인가 로직 추가 예정
+        validatePostOwner(memberId, findPost);
 
         findPost.update(requestDto.getContent(), requestDto.getImageUrl());
 
@@ -78,14 +75,31 @@ public class PostServiceImpl implements PostService {
 
     @Transactional
     @Override
-    public void deletePost(Long postId) {
-        Post findPost = postRepository.findById(postId)
-                .orElseThrow(() -> new PostNotFoundException());
+    public void deletePost(Long memberId, Long postId) {
+        Post findPost = findPostByIdOrElseThrow(postId);
 
         // TODO: 추후에 작성자를 확인하는 인가 로직 추가 예정
+        validatePostOwner(memberId, findPost);
 
         postRepository.delete(findPost);
     }
 
+    private Member findMemberByIdOrElseThrow(Long memberId) {
+        // FIXME: 현재는 임시로 ResponseStatusException 사용 중.
+        // 추후 member 도메인에서 커스텀 예외(MemberNotFoundException 등) 정의되면 교체할 것.
+        return memberRepository.findById(memberId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+    }
+
+    private Post findPostByIdOrElseThrow(Long postId) {
+        return postRepository.findById(postId)
+                .orElseThrow(PostNotFoundException::new);
+    }
+
+    private static void validatePostOwner(Long memberId, Post findPost) {
+        if (!findPost.getMember().getId().equals(memberId)) {
+            throw new PostUnauthorizedException();
+        }
+    }
 
 }
